demo/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
blinker==1.5
cachetools
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
datasets==2.12.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.1
grpcio-status==1.48.2
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
langchain==0.0.171
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.2
openai==0.27.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
regex==2023.5.5
requests==2.28.2
responses==0.18.0
rich==13.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.13
stack-data==0.6.2
streamlit==1.25.0
streamlit-nested-layout==0.1.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.29.2
typing_extensions==4.4.0
typing-inspect==0.8.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wcwidth==0.2.6
wheel==0.37.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.12.1
streamlit-chat==0.0.2.2
pinecone-client==2.2.2
torch==2.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
mlflow==2.4.1
pinecone-text==0.5.2
rank-bm25==0.2.2
faiss-cpu
faiss-gpu